requirements.txt
chromadb
docx2txt
edge-tts
gradio
huggingface_hub
llama-index
llama-index-core
llama-index-embeddings-huggingface
llama-index-llms-huggingface-api
llama-index-llms-mistralai
llama-index-vector-stores-chroma
python-dotenv
SpeechRecognition